centralize fixtures
